ajoute les promos sur les articles

+++ db/migrate/20190424121712_add_discount_to_stock.rb
@@ -1,5 +1,5 @@
 class AddDiscountToStock < ActiveRecord::Migration[5.2]
   def change
-  	add_column :stocks, :discount, :integer, default: 0
+  	add_column :stocks, :discount, :decimal, default: 0
   end
 end
